gcp_cloud_functions/ftp_xml_df_bq_steaming/requirements.txt
-i https://pypi.org/simple
google-cloud-storage==1.36.2
google-cloud-bigquery ==2.11.0
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-api-core==1.26.1
requests==2.22.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
google-auth==1.27.1
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.8
protobuf==3.12.0
pyasn1-modules==0.2.5
pyasn1==0.4.5
pytz==2019.1
rsa==4.7.2
six==1.14.0
urllib3==1.25.3
lxml==4.6.3
pandas==1.2.3
fsspec==0.8.7
gcsfs==0.7.2
pandas-gbq ==0.14.1

